fixed bits and pieces in vm

+++ output.c
@@ -10,12 +10,12 @@
 #include "memory.h"
 
 // convert a list into a string in "..., ..., ...]" form, yes without the first bracket
-char *list_to_str(int *list, int size) {
+char *list_to_str(long long *list, long long size) {
 
     char *s = (char*)malloc(MAX_MSG_LEN * sizeof(char));
     strcpy(s, "");
 
-    for (int i = 0; i < size; i++) {
+    for (long long i = 0; i < size; i++) {
         char addr_str[ADDR_STR_LEN];
         if ((i != 0) && (list[i] == 0)) {
             // initialisation caused some 0 at the end
@@ -23,7 +23,7 @@ char *list_to_str(int *list, int size) {
             s[strlen(s) - 1] = ']';  // replace the last comma with ]
             break;
         }
-        sprintf(addr_str, i < size - 1 ? "%d," : "%d]", list[i]);
+        sprintf(addr_str, i < size - 1 ? "%lld," : "%lld]", list[i]);
         strcat(s, addr_str);
     }
 
@@ -31,28 +31,28 @@ char *list_to_str(int *list, int size) {
 }
 
 // print out the status when changed
-void print_status(int time, int status, int proc_id, char *msg) {
+void print_status(long long time, int status, long long proc_id, char *msg) {
     switch (status) {
         case RUNNING:
-            printf("%d, RUNNING, id=%d, %s\n", time, proc_id, msg);
+            printf("%lld, RUNNING, id=%lld, %s\n", time, proc_id, msg);
             break;
         case FINISHED:
-            printf("%d, FINISHED, id=%d, %s\n", time, proc_id, msg);
+            printf("%lld, FINISHED, id=%lld, %s\n", time, proc_id, msg);
             break;
         case EVICTED:
-            printf("%d, EVICTED, %s\n", time, msg);
+            printf("%lld, EVICTED, %s\n", time, msg);
         default:
             break;
     }
 }
 
 // print stats at the end
-void print_stats(int process_executed, int total_turnaround, float total_overhead, float max_overhead, int time,
-                 int avg_throughput, int min_throughput, int max_throughput) {
-    printf("Throughput %d, %d, %d\n", avg_throughput, min_throughput, max_throughput);
-    printf("Turnaround time %d\n", (int)ceil((double)total_turnaround / (double)process_executed));
+void print_stats(long long process_executed, long long total_turnaround, float total_overhead, float max_overhead, long long time,
+                 long long avg_throughput, long long min_throughput, long long max_throughput) {
+    printf("Throughput %lld, %lld, %lld\n", avg_throughput, min_throughput, max_throughput);
+    printf("Turnaround time %lld\n", (long long)ceil((double)total_turnaround / (double)process_executed));
     printf("Time overhead %.2f %.2f\n", max_overhead, total_overhead / process_executed);
-    printf("Makespan %d\n", time);
+    printf("Makespan %lld\n", time);
 }
 
 // print all the specs given by command line
@@ -60,8 +60,8 @@ void print_spec(char* file_name, spec_t spec) {
 
     int sch_algo = spec.sch_algo;
     int mem_allo = spec.mem_allo;
-    int mem_size = spec.mem_size;
-    int quantum = spec.quantum;
+    long long mem_size = spec.mem_size;
+    long long quantum = spec.quantum;
 
     char sch_algo_name[MAX_NAME_LEN], mem_allo_name[MAX_NAME_LEN];
 
@@ -88,8 +88,8 @@ void print_spec(char* file_name, spec_t spec) {
     printf("Filename:             %s\n", file_name);
     printf("Scheduling Algorithm: %s\n", sch_algo_name);
     printf("Memory Allocation:    %s\n", mem_allo_name);
-    printf("Memory Size:          %d\n", mem_size);
-    printf("Quantum:              %d\n", quantum);
+    printf("Memory Size:          %lld\n", mem_size);
+    printf("Quantum:              %lld\n", quantum);
 
 }
 
@@ -99,11 +99,11 @@ void print_process(process_t *process) {
         perror("ERROR printing null process");
         exit(EXIT_FAILURE);
     }
-    printf("===== PROCESS %2d =====\n", process->id);
-    printf("  Arrival Time: %d\n", process->arrival_time);
-    printf("  Job Time:     %d\n", process->job_time);
-    printf("  Mem Required: %d\n", process->mem_req);
-    printf("  Remain Time:  %d\n\n", process->remaining_time);
+    printf("===== PROCESS %2lld =====\n", process->id);
+    printf("  Arrival Time: %lld\n", process->arrival_time);
+    printf("  Job Time:     %lld\n", process->job_time);
+    printf("  Mem Required: %lld\n", process->mem_req);
+    printf("  Remain Time:  %lld\n\n", process->remaining_time);
 }
 
 // print the whole list of processes
@@ -122,14 +122,13 @@ void print_process_list(process_list_t *process_list) {
 }
 
 // print information of a unit of memory
-void print_memory(unit_t mem_unit, int mem_id) {
-    printf("%2d. === proc_id:   %d\n", mem_id, mem_unit.proc_id);
-    // printf("    === time_used: %d\n", mem_unit.time_used);
+void print_memory(unit_t mem_unit, long long mem_id) {
+    printf("%2lld. === proc_id:   %lld\n", mem_id, mem_unit.proc_id);
 }
 
 // print list of memory
-void print_memory_list(unit_t *mem_list, int size) {
-    for (int i = 0; i < size; i++) {
+void print_memory_list(unit_t *mem_list, long long size) {
+    for (long long i = 0; i < size; i++) {
         print_memory(mem_list[i], i);
     }
 }
@@ -141,8 +140,8 @@ void print_status_node(status_t *status) {
         exit(EXIT_FAILURE);
     }
     printf("===== STATUS %2d =====\n", status->status);
-    printf("  Start:  %d\n", status->start);
-    printf("  Size:   %d\n", status->size);
+    printf("  Start:  %lld\n", status->start);
+    printf("  Size:   %lld\n", status->size);
 }
 
 // print the whole list of status
